Moved altitude Kalman filter to its own files
