+++ Inc/AltitudeKalmanFilter.h
@@ -1,0 +1,8 @@
+#pragma once
+
+struct KalmanStateVector
+{
+    double altitude;
+    double velocity;
+    double acceleration;
+};
